Fixed mutual dependencies resolution
I marked packages as `processed` so that I do not enqueue them multiple
times. When we encounter package that is already processed we early-quit
from the function before enqueueing logic kicks in.

Unfortunately the early-exit was done too early, without considering
connections that led to this package. This resulted in a bug
https://github.com/anvaka/npmgraph.an/issues/11

If the dep graph looks like:
A->C@1.0.0
B->C@1.0.0

C will be shown as only the child of one of either A or B, not both.

This change fixes it by quiting before enqueueing logic, but after links
are constructed